Fix tests, and use granularity in non inclusive relations
