+++ generator/src/main/java/com/scottlogic/deg/generator/fieldspecs/relations/BeforeDateRelation.java
@@ -17,35 +17,53 @@
 package com.scottlogic.deg.generator.fieldspecs.relations;
 
 import com.scottlogic.deg.common.profile.Field;
+import com.scottlogic.deg.generator.fieldspecs.FieldSpec;
 import com.scottlogic.deg.generator.restrictions.linear.Limit;
 import com.scottlogic.deg.generator.restrictions.linear.LinearRestrictions;
 import com.scottlogic.deg.generator.restrictions.linear.LinearRestrictionsFactory;
 
 import java.time.OffsetDateTime;
 
+import static com.scottlogic.deg.common.util.Defaults.ISO_MAX_DATE;
+import static com.scottlogic.deg.common.util.Defaults.ISO_MIN_DATE;
 import static com.scottlogic.deg.generator.utils.Defaults.DATETIME_MAX_LIMIT;
 import static com.scottlogic.deg.generator.utils.Defaults.DATETIME_MIN_LIMIT;
 
-public class BeforeDateRelation extends AbstractDateInequalityRelation {
+public class BeforeDateRelation implements FieldSpecRelations {
+    private final Field main;
+    private final Field other;
     private final boolean inclusive;
 
     public BeforeDateRelation(Field main, Field other, boolean inclusive) {
-        super(main, other);
+        this.main = main;
+        this.other = other;
         this.inclusive = inclusive;
     }
 
     @Override
-    public OffsetDateTime getRelevantValue(LinearRestrictions<OffsetDateTime> restrictions) {
-        if (restrictions != null) {
-            return restrictions.getMax();
-        } else {
-            return null;
+    public FieldSpec reduceToRelatedFieldSpec(FieldSpec otherValue) {
+        LinearRestrictions<OffsetDateTime> lr = (LinearRestrictions) otherValue.getRestrictions();
+        if (lr == null){
+            return FieldSpec.empty();
         }
+
+        OffsetDateTime max = lr.getMax();
+        if (!inclusive){
+            max = lr.getGranularity().getPrevious(max);
+        }
+
+        return FieldSpec.fromRestriction(new LinearRestrictions<>(ISO_MIN_DATE, max, lr.getGranularity()));
+    }
+
+
+    @Override
+    public Field main() {
+        return main;
     }
 
     @Override
-    protected LinearRestrictions<OffsetDateTime> appendValueToRestrictions(OffsetDateTime value) {
-        return LinearRestrictionsFactory.createDateTimeRestrictions(DATETIME_MIN_LIMIT, new Limit<>(value, inclusive));
+    public Field other() {
+        return other;
     }
 
     @Override
